refactor: modularize training code by moving utilities to training_utils.py
